% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Bind 2 factors}
\usage{
fbind(a, b)
}
\arguments{
\item{a}{facotr}

\item{b}{factor}
}
\value{
factor
}
\description{
Bind 2 factors
}
\examples{
fbind(factor("sb"), factor("hh"))
}
